Move all JS files to a "src" directory
This is groundwork for the Typescript conversion. It will allow easily
building a clean Docker image for prod. We'll transpile TS to a
separate output directory called `build` or whatever, and only include
that in the final image. This change makes it so that all the code
assumes it's one level deeper in the tree.

To test, I:

- Ran `yarn test`
- Ran `yarn dev` and made requests locally
- Built the Docker image, ran it locally, and made requests
