+++ src/lib/incentives-calculation.js
