+++ src/plugins/sensible.js
