+++ src/plugins/swagger.js
